refactor: fsd 구조 변환
