+++ apps/web/src/features/auth/ui/index.ts
@@ -1,0 +1,1 @@
+export { LoginBottomSheet } from './loginBottomSheet';
