+++ apps/web/src/features/auth/ui/loginBottomSheet/LoginBottomSheet.stories.tsx
@@ -1,0 +1,52 @@
+import type { Meta, StoryObj } from '@storybook/nextjs';
+import { useState } from 'react';
+import { BottomSheet } from '@/shared/ui/bottomSheet';
+import { Button } from '@/shared/ui';
+import { LoginBottomSheet } from './LoginBottomSheet';
+
+const meta: Meta<typeof LoginBottomSheet> = {
+  title: 'Components/BottomSheet/LoginBottomSheet',
+  component: LoginBottomSheet,
+  parameters: {
+    layout: 'fullscreen',
+  },
+};
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Default: Story = {
+  render: () => {
+    const TemplateDemo = () => {
+      const [isOpen, setIsOpen] = useState(false);
+
+      return (
+        <div
+          style={{
+            display: 'flex',
+            justifyContent: 'center',
+            alignItems: 'center',
+            height: '100vh',
+          }}>
+          <Button variant='brand' onClick={() => setIsOpen(true)}>
+            로그인 열기
+          </Button>
+
+          <BottomSheet isOpen={isOpen} onClose={() => setIsOpen(false)}>
+            <LoginBottomSheet />
+          </BottomSheet>
+        </div>
+      );
+    };
+
+    return <TemplateDemo />;
+  },
+  parameters: {
+    docs: {
+      story: {
+        inline: false,
+        height: '500px',
+      },
+    },
+  },
+};
