+++ apps/web/src/features/auth/ui/loginBottomSheet/index.ts
@@ -1,0 +1,1 @@
+export { LoginBottomSheet } from './LoginBottomSheet';
