+++ apps/web/src/features/expense/index.ts
@@ -1,0 +1,16 @@
+export {
+  ExpenseFormBottomSheet,
+  CategoryBottomSheetTemplate,
+  CalendarBottomSheetTemplate,
+  DatePickerBottomSheetTemplate,
+  IconPickerBottomSheetTemplate,
+} from './ui';
+
+export type {
+  ExpenseFormBottomSheetProps,
+  Category,
+  CategoryBottomSheetTemplateProps,
+  CalendarBottomSheetTemplateProps,
+  DatePickerBottomSheetTemplateProps,
+  IconPickerBottomSheetTemplateProps,
+} from './ui';
