+++ apps/web/src/features/expense/ui/expenseBottomSheet/CalendarBottomSheet.stories.tsx
@@ -1,0 +1,58 @@
+import type { Meta, StoryObj } from '@storybook/nextjs';
+import { useState } from 'react';
+import { BottomSheet } from '@/shared/ui/bottomSheet';
+import { Button } from '@/shared/ui';
+import { CalendarBottomSheetTemplate } from './CalendarBottomSheet';
+
+const meta: Meta<typeof CalendarBottomSheetTemplate> = {
+  title: 'Components/BottomSheet/CalendarBottomSheet',
+  component: CalendarBottomSheetTemplate,
+  parameters: {
+    layout: 'fullscreen',
+  },
+};
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Default: Story = {
+  render: () => {
+    const TemplateDemo = () => {
+      const [isOpen, setIsOpen] = useState(false);
+
+      return (
+        <div
+          style={{
+            display: 'flex',
+            justifyContent: 'center',
+            alignItems: 'center',
+            height: '100vh',
+          }}>
+          <Button variant='brand' onClick={() => setIsOpen(true)}>
+            소비일 수정 열기
+          </Button>
+
+          <BottomSheet isOpen={isOpen} onClose={() => setIsOpen(false)}>
+            <CalendarBottomSheetTemplate
+              onClose={() => setIsOpen(false)}
+              onConfirm={() => {
+                alert('소비일 수정 완료');
+                setIsOpen(false);
+              }}
+            />
+          </BottomSheet>
+        </div>
+      );
+    };
+
+    return <TemplateDemo />;
+  },
+  parameters: {
+    docs: {
+      story: {
+        inline: false,
+        height: '600px',
+      },
+    },
+  },
+};
