+++ apps/web/src/features/expense/ui/expenseBottomSheet/DatePickerBottomSheet.tsx
@@ -1,8 +1,7 @@
-// CalendarBottomSheetTemplate
-import BaseBottomSheetTemplate from './BaseBottomSheetTemplate';
+import BaseBottomSheetTemplate from '../../../../shared/ui/bottomSheet/templates/BaseBottomSheetTemplate';
 import React from 'react';
 
-export interface CalendarBottomSheetTemplateProps {
+export interface DatePickerBottomSheetTemplateProps {
   /** 선택된 날짜 */
   selectedDate?: Date;
   /** 날짜 선택 시 콜백 */
@@ -13,13 +12,13 @@ export interface CalendarBottomSheetTemplateProps {
   onClose?: () => void;
 }
 
-export const CalendarBottomSheetTemplate = ({
+export const DatePickerBottomSheetTemplate = ({
   onConfirm,
   onClose,
-}: CalendarBottomSheetTemplateProps) => {
+}: DatePickerBottomSheetTemplateProps) => {
   return (
     <BaseBottomSheetTemplate>
-      <BaseBottomSheetTemplate.Header type='close' text='소비일 수정' onClose={onClose} />
+      <BaseBottomSheetTemplate.Header type='close' text='월 선택' onClose={onClose} />
       <div>{/* TODO: 달력 컴포넌트 */}</div>
       <BaseBottomSheetTemplate.Button label='선택' onClick={onConfirm} />
     </BaseBottomSheetTemplate>
