+++ apps/web/src/features/expense/ui/expenseBottomSheet/ExpenseFormBottomSheet.css.ts
@@ -1,5 +1,5 @@
 import { style } from '@vanilla-extract/css';
-import { radius, spacing, vars } from '../../theme.css';
+import { radius, spacing, vars } from '@/shared/ui/theme.css';
 
 export const inputWrapper = style({
   marginBottom: '2.1rem',
