+++ apps/web/src/features/expense/ui/expenseBottomSheet/index.ts
@@ -1,0 +1,14 @@
+export { ExpenseFormBottomSheet } from './ExpenseFormBottomSheet';
+export type { ExpenseFormBottomSheetProps, Category } from './ExpenseFormBottomSheet';
+
+export { CategoryBottomSheetTemplate } from './CategoryBottomSheet';
+export type { CategoryBottomSheetTemplateProps } from './CategoryBottomSheet';
+
+export { CalendarBottomSheetTemplate } from './CalendarBottomSheet';
+export type { CalendarBottomSheetTemplateProps } from './CalendarBottomSheet';
+
+export { DatePickerBottomSheetTemplate } from './DatePickerBottomSheet';
+export type { DatePickerBottomSheetTemplateProps } from './DatePickerBottomSheet';
+
+export { IconPickerBottomSheetTemplate } from './IconPickerBottomSheet';
+export type { IconPickerBottomSheetTemplateProps } from './IconPickerBottomSheet';
