+++ apps/web/src/features/expense/ui/index.ts
@@ -1,0 +1,16 @@
+export {
+  ExpenseFormBottomSheet,
+  CategoryBottomSheetTemplate,
+  CalendarBottomSheetTemplate,
+  DatePickerBottomSheetTemplate,
+  IconPickerBottomSheetTemplate,
+} from './expenseBottomSheet';
+
+export type {
+  ExpenseFormBottomSheetProps,
+  Category,
+  CategoryBottomSheetTemplateProps,
+  CalendarBottomSheetTemplateProps,
+  DatePickerBottomSheetTemplateProps,
+  IconPickerBottomSheetTemplateProps,
+} from './expenseBottomSheet';
