+++ apps/web/src/shared/ui/bottomSheet/templates/index.ts
@@ -1,0 +1,1 @@
+export { default as BaseBottomSheetTemplate } from './BaseBottomSheetTemplate';
